+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
--   Fixed `TabProvider` setting the `activeTabIndex` to `-1` on unmount.
+-   Fixed tabs and progress tracker active index synchronization.
+
 ## [1.0.11][] - 2021-02-26
 
 ### Added